Fix client building
[#172225423]
