+++ functions/src/helpers.ts
@@ -1,19 +1,6 @@
-import { BaseResource, ReadWriteTokenAccessRule, ResourceType } from "./resource-types";
+import { ResourceType } from "./resource-types";
 import { FireStoreResourceSettings } from "./firestore-types";
 import { BaseResourceObject } from "./base-resource-object";
-
-export const getRWTokenFromAccessRules = (resource: BaseResource): string | undefined => {
-  if (!resource.accessRules) {
-    return undefined;
-  }
-  const readWriteTokenRules = resource.accessRules.filter(r => r.type === "readWriteToken");
-  if (readWriteTokenRules.length > 0) {
-    // There's no reason to have more than one readWriteToken rules (so in fact multiple read write tokens).
-    // But even if that happens, it's enough to get any of them to have access to the resource.
-    return (readWriteTokenRules[0] as ReadWriteTokenAccessRule).readWriteToken;
-  }
-  return undefined;
-};
 
 export const makeCachedSettingsGetter = (fsDb: FirebaseFirestore.Firestore, env: string) => {
   const settingsCache: {[key: string]: FireStoreResourceSettings} = {};
